DataEnggProjects/requirements.txt
pandas
google-cloud-storage
numpy
google-cloud
requests
google-cloud-bigquery
